requirements.txt
streamlit
requests
openai
azure-ai-search
azure-identity
dotenv
python-dotenv
azure-search-documents
azure-core